scripts/requirements.txt
crcmod==1.7
hexdump==3.3
Pillow==11.0.0
